{"schema":"decodex/commit/1","summary":"Modularize state runtime records","authority":"manual"}
